refactor: Move commands into subdirectories
That's makes it easier to consolidate command specific modules.
