+++ lib/buildEjectCommand/index.js
@@ -2,12 +2,9 @@
  * @flow
  */
 
-import type {MakeItem, MakeRawItem, MakeDefine, MakeRule} from './Makefile';
-import type {
-  Sandbox,
-  PackageInfo
-} from './Sandbox';
-import type {EnvironmentGroup} from './PackageEnvironment';
+import type {MakeItem, MakeRawItem, MakeDefine, MakeRule} from '../Makefile';
+import type {Sandbox, PackageInfo} from '../Sandbox';
+import type {EnvironmentGroup} from '../PackageEnvironment';
 
 const mkdirp = require('mkdirp').sync;
 const childProcess = require('child_process');
@@ -15,17 +12,17 @@ const fs = require('fs');
 const crypto  = require('crypto');
 const path = require('path');
 const outdent = require('outdent');
-const {flattenArray} = require('./Utility');
+const {flattenArray} = require('../Utility');
 
-const RUNTIME = fs.readFileSync(require.resolve('./esyBuildRuntime.sh'), 'utf8');
+const RUNTIME = fs.readFileSync(require.resolve('./runtime.sh'), 'utf8');
 
 const {
   traversePackageDependencyTree,
   collectTransitiveDependencies,
   packageInfoKey,
-} = require('./Sandbox');
-const PackageEnvironment = require('./PackageEnvironment');
-const Makefile = require('./Makefile');
+} = require('../Sandbox');
+const PackageEnvironment = require('../PackageEnvironment');
+const Makefile = require('../Makefile');
 
 const curWorkingDirectory = process.cwd();
