+++ lib/installCommand/index.js
@@ -2,7 +2,7 @@
  * @flow
  */
 
-import type {Sandbox} from './Sandbox';
+import type {Sandbox} from '../Sandbox';
 
 import {tmpdir} from 'os';
 import path from 'path';
@@ -15,7 +15,7 @@ import tempfile from 'tempfile';
 import * as pnpm from '@esy-ocaml/pnpm';
 import mkdirp from '@esy-ocaml/pnpm/lib/fs/mkdirp';
 import {fetchFromTarball} from '@esy-ocaml/pnpm/lib/install/fetchResolution';
-import {hash} from './Utility';
+import {hash} from '../Utility';
 
 export type PackageJsonCollection = {
   versions: {
@@ -39,7 +39,7 @@ export type OpamInfo = {
   }
 };
 
-const OPAM_METADATA_STORE = path.join(__dirname, '..', 'opam-packages');
+const OPAM_METADATA_STORE = path.join(__dirname, '..', '..', 'opam-packages');
 
 const USER_HOME: string = (process.env.HOME: any)
