+++ package.json
@@ -16,10 +16,13 @@
     "babel-preset-env": "^1.1.8",
     "babel-register": "^6.22.0",
     "cli-color": "^1.1.0",
+    "execa": "^0.6.0",
     "mkdirp": "^0.5.1",
+    "mz": "^2.6.0",
     "outdent": "^0.3.0",
     "path-is-inside": "^1.0.2",
     "resolve": "^1.1.7",
+    "semver": "^5.3.0",
     "tempfile": "^1.1.1"
   },
   "repository": {
